requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
python-multipart==0.0.6
faker==18.13.0
requests==2.31.0
python-dotenv==0.21.1
